requirements.txt
numpy==1.26.4
pandas==2.3.3
scikit-learn==1.7.2
scikit-surprise==1.1.4
mlflow==3.6.0
pydantic==2.12.4
python-dateutil==2.9.0.post0
tqdm==4.67.1
